+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Added `Session::regenerateId()` method to properly prevent session fixation issues
 1. [](#bugfix)
     * Fixed `validation: strict` not working in blueprints [#1273](https://github.com/getgrav/grav/issues/1273)
+    * Fixed Flex Pages having broken `isFirst()`, `isLast()`, `prevSibling()`, `nextSibling()` and `adjacentSibling()`
 
 # v1.7.0-rc.6
 ## 02/11/2020